@@ -1,18 +1,15 @@
 package com.example.demo.controllers;
 
+import com.example.demo.model.RequestObj;
+import com.example.demo.model.ResponseObj;
+import com.example.demo.util.RestUtil;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
-import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.example.demo.model.RequestObj;
-import com.example.demo.model.ResponseObj;
-import com.example.demo.util.RestUtil;
-
-import lombok.extern.slf4j.Slf4j;
 
 /**
  * The Class MyRestController.
